requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.10.0
django-extensions==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.3.0
django-utils
django-utils2
